+++ src/components/QuickstartTile.js
@@ -21,6 +21,7 @@ const QuickstartTile = ({
   title,
   name,
   logo,
+  logoSvg,
   level,
   className,
   summary,
@@ -64,6 +65,7 @@ const QuickstartTile = ({
       css={css`
         --tile-image-height: 100px; /* Logo image height */
         --title-row-height: auto; /* Title height to allow space for longer string */
+        --tile-content-height: 180px;
         padding: 0 22px 35px 24px;
         overflow: hidden;
         height: 360px;
@@ -95,7 +97,7 @@ const QuickstartTile = ({
         display: grid;
         align-items: flex-start;
         grid-gap: 0.2rem;
-        grid-template-rows: var(--tile-image-height) 152px auto;
+        grid-template-rows: var(--tile-image-height) var(--tile-content-height) auto;
         grid-template-columns: auto;
         grid-template-areas:
           'logo logo'
@@ -121,6 +123,7 @@ const QuickstartTile = ({
         >
           <QuickstartImg
             imageNode={logo}
+            svgNode={logoSvg}
             packName={title || name}
             css={css`
               object-fit: scale-down;
@@ -141,12 +144,14 @@ const QuickstartTile = ({
       <div
         css={css`
           grid-area: text;
-          overflow-wrap: anywhere;
+          height: var(--tile-content-height);
+          overflow: hidden;
+          display: flex;
+          flex-flow: column wrap;
         `}
       >
         <h4
           css={css`
-            grid-area: title;
             font-weight: normal;
             font-size: 24px;
             letter-spacing: -0.5%;
@@ -156,12 +161,7 @@ const QuickstartTile = ({
           {SHIELD_LEVELS.includes(level) && <Icon name="nr-check-shield" />}
         </h4>
 
-        <div
-          css={css`
-            grid-area: summary;
-            align-self: flex-start;
-          `}
-        >
+        <div>
           <p
             css={css`
               font-size: 18px;
@@ -242,6 +242,7 @@ QuickstartTile.propTypes = {
   title: PropTypes.string.isRequired,
   slug: PropTypes.string.isRequired,
   logo: PropTypes.object,
+  logoSvg: PropTypes.object,
   summary: PropTypes.string,
   level: PropTypes.string,
   className: PropTypes.string,